streamlit/requirements.txt
streamlit
pandas
plotly
pillow
numpy
scikit-learn
nltk
transformers
transformer-utils
scipy
torch